house_prediction/requirements.txt
kfp==2.7.0
google-cloud-aiplatform==1.42.1
google-cloud-storage==2.14.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gcsfs==2024.2.0
numpy>=1.22.4,<2
pandas>=2.2.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.0